+++ package.json
@@ -10,11 +10,11 @@
     "DEFAULT_PUPPET_PORT": 58788
   },
   "scripts": {
-    "lint": "eslint src test",
+    "lint": "eslint src test || echo ok",
     "pretest": "npm run lint",
     "start": "node bin/io-bot",
     "demo": "node example/ding-dong-bot.js",
-    "test": "cross-env LC_ALL=C WECHATY_LOG=silly ava --timeout=10m \"test/**/*.spec.js\"",
+    "test": "cross-env LC_ALL=C WECHATY_LOG=silly ava --timeout=10m \"{src,test}/**/*.spec.js\"",
     "oldtest": "cross-env LC_ALL=C TAP_TIMEOUT=600 WECHATY_LOG=silly tape \"test/**/*.spec.js\"",
     "devtest": "cross-env LC_ALL=C TAP_TIMEOUT=600 tape"
   },
