+++ test/electron.spec.js
@@ -3,3 +3,5 @@ import { test } from 'ava'
 test('Electron smoke testing', async t => {
   t.true(true, 'test')
 })
+
+test.todo('Electron open wx')
